+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.1.10-alpha
 - Add support for PHP7 runtime errors to be handled in the same way as Exceptions
 - Implement NotFoundExceptionHandler and pass through all NotFoundExceptions to this handler.
+- Factor out redirects in AccountController into a separate Redirector service (#680)
 - Fix some PSR-2 compliance issues
 
 ## v4.1.9-alpha